Add API for releasing zero-copy USB buffers.
API functions are added to provide for the use of zero-copy buffers
beyond the lifetime of the async callback.  This allows the number of
calls to the callback to be rapidly tracked in order to measure and
predict latency and data loss, even when processing the data is slower
than receiving it, without requiring memcpy() be used to take data out
of the buffers.  This addition benefits mostly multithreaded clients.

rtlsdr_read_async_ex is added which passes a buffer handle to the
callback.  Additionally, it calls its callback even in the case of
failed bulk transfers, to inform of possible gaps in the stream.

rtlsdr_read_async_ex_release is added, which resubmits buffers to libusb
when client code has finished with them.

rtlsdr_read_async is rewritten in terms of the more general
rtlsdr_read_async_ex.

Significant documentation is added to explain how these functions relate
to latency.